pyproject.toml
[project]
name = "sae-ppa-chatbot"
package-mode = false
version = "0.1.0"
description = ""
authors = [
    {name = "arnoc-intuvision",email = "arnoc@intuvision.co.za"}
]
readme = "README.md"

requires-python = ">=3.11,<4.0"
dependencies = [
    "boto3 (>=1.40.12,<2.0.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-genai (>=1.31.0,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"